services/execution/requirements.txt
fastapi==0.115.4
uvicorn==0.34.0
boto3==1.35.0
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
pydantic-settings==2.5.2

